pyproject.toml
[tool.poetry]
name = "discord-prank"
version = "0.1.0"
description = ""
authors = ["amirmaula1 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.2.0"
Flask = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
